Backend/src/requirements.txt
Werkzeug
flask
flask-login
flask-cors
mysql-connector
mysql-connector-python
flask_sqlalchemy
pymysql
cryptography
python-dotenv